fixed all bugs on matrix before finishing and adding big O
